lab1_context_state/pyproject.toml
[project]
name = "lab1-context-state"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.19.0",
]
